更新README说明：增加1.1.0版本新增加的功能和注意点 修改测试代码 解决一个查询所有模型的一个bug

+++ Tests/TestEasyFMDB/TestEasyFMDBTests/ZyxContact.h
@@ -11,7 +11,7 @@
 @interface ZyxContact : ZyxBaseModel
 
 @property (nonatomic, copy) NSString *name;
-@property (nonatomic, assign) int age;
+@property (nonatomic, assign) NSInteger age;
 @property (nonatomic, copy) NSString *homeAddress;
 @property (nonatomic, copy) NSString *workAddress;
 @property (nonatomic, copy) NSString *mobilePhone;
